Refactor to get Pwr2 as an umbrella app
