+++ apps/webapp/web/static/js/star_wars/components/StarWarsShip.js
