+++ apps/webapp/web/templates/page/graphiql.html.eex
